启用message_sender.py
